requirements.txt
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
distlib==0.3.3
filelock==3.3.1
Flask==2.0.2
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.2
pandas==1.3.4
platformdirs==2.4.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.25
urllib3==1.26.7
virtualenv==20.8.1
Werkzeug==2.0.2
WTForms==3.0.0
